feat: get all assignments and filter assignments
