+++ fe/src/features/assignments/api/get-assignments-with-filter.tsx
@@ -1,0 +1,10 @@
+import type { FilterData } from '../get-all-assignments/model'
+import { httpClient } from '@/lib'
+
+export const getAssignmentsWithFilter = async (values: FilterData) => {
+  return await httpClient.get('/assignments/filter', {
+    params: {
+      ...values,
+    },
+  })
+}
