+++ fe/src/features/assignments/api/get-assignments.tsx
@@ -1,10 +1,5 @@
 import { httpClient } from '@/lib'
-import type { AssignmentsResponse } from '../get-all-assignments/model'
 
-export const getAssignments = async (value?: AssignmentsResponse) => {
-  return await httpClient.get('/assignments/filter', {
-    params: {
-      ...value,
-    },
-  })
+export const getAssignments = async () => {
+  return await httpClient.get('/assignments/filter')
 }
