+++ fe/src/features/assignments/api/index.ts
@@ -1,1 +1,2 @@
 export * from './get-assignments'
+export * from './get-assignments-with-filter'
