+++ fe/src/features/assignments/get-all-assignments/list-assignments.tsx
@@ -1,17 +1,13 @@
-import { useEffect, useState, useTransition } from 'react'
-import { Card, CardHeader, CardTitle, CardContent, CardDescription } from '@/components/ui/card'
-import { DataTable, Button } from '@/components/ui'
-import { Skeleton } from '@/components/ui/skeleton'
-import { ClipboardList, Search } from 'lucide-react'
-import { Input } from '@/components/ui/input'
-import type { AssignmentsResponse } from './model/type'
-import { getData, tryCatch } from '@/utils'
-import { getAssignments } from '../api'
+import { useEffect, useState } from 'react'
+import { Card, CardHeader, CardTitle, CardContent, CardDescription, Skeleton, DataTable } from '@/components/ui'
+import { ClipboardList } from 'lucide-react'
+import type { AssignmentsResponse } from './model'
+import { getData } from '@/utils'
+import { getAssignments, getAssignmentsWithFilter } from '../api'
 import { columnsAssignmentsTable } from './column-table'
 import { useDebounce } from '@/hooks'
-import { FilterX } from 'lucide-react'
 import { useSearchParams } from 'react-router-dom'
-import { toast } from 'sonner'
+import { ButtonClearFilter, AssetNameFilter, AssignedToFilter, AssignedByFilter } from './_components'
 
 const ListAssignments = () => {
   const [searchParam, setSearchParam] = useSearchParams()
@@ -22,30 +18,17 @@ const ListAssignments = () => {
     emailAssign: '',
     emailAssigned: '',
   })
-  const resetFilters = () => {
-    setFilters({
-      assetName: '',
-      emailAssign: '',
-      emailAssigned: '',
-    })
-  }
 
   const filterData = useDebounce(filters, 1000)
 
   const fetchData = async () => {
     setIsLoading(true)
-    const data = await getData(getAssignments, setAssignments)
-    console.log('Assignments data:', data)
+    await getData(getAssignments, setAssignments)
     setIsLoading(false)
   }
   const getAssetsFilterData = async () => {
     setIsLoading(true)
-    const { data, error } = await tryCatch(getAssignments(filterData))
-    if (error) {
-      toast.error(error?.message || 'Failed to load assets')
-      return
-    }
-    setAssignments(data.data)
+    await getData(() => getAssignmentsWithFilter(filterData), setAssignments)
     setIsLoading(false)
   }
 
@@ -88,51 +71,22 @@ const ListAssignments = () => {
       <CardContent>
         <div className='space-y-4'>
           <div className='flex flex-col gap-4 md:flex-row'>
-            <div className='flex-1 space-y-2'>
-              <label className='text-sm font-medium'>Asset Name</label>
-              <div className='relative'>
-                <Search className='text-muted-foreground absolute top-2.5 left-2 h-4 w-4' />
-                <Input
-                  placeholder='Filter by asset name...'
-                  className='pl-8'
-                  value={filters.assetName}
-                  onChange={(e) => setFilters((prev) => ({ ...prev, assetName: e.target.value }))}
-                />
-              </div>
-            </div>
-            <div className='flex-1 space-y-2'>
-              <label className='text-sm font-medium'>Assigned By (Email)</label>
-              <div className='relative'>
-                <Search className='text-muted-foreground absolute top-2.5 left-2 h-4 w-4' />
-                <Input
-                  placeholder='Filter by assigner email...'
-                  className='pl-8'
-                  value={filters.emailAssign}
-                  onChange={(e) => setFilters((prev) => ({ ...prev, emailAssign: e.target.value }))}
-                />
-              </div>
-            </div>
-            <div className='flex-1 space-y-2'>
-              <label className='text-sm font-medium'>Assigned To (Email)</label>
-              <div className='relative'>
-                <Search className='text-muted-foreground absolute top-2.5 left-2 h-4 w-4' />
-                <Input
-                  placeholder='Filter by assignee email...'
-                  className='pl-8'
-                  value={filters.emailAssigned}
-                  onChange={(e) => setFilters((prev) => ({ ...prev, emailAssigned: e.target.value }))}
-                />
-              </div>
-            </div>
-            <Button
-              disabled={!filters.assetName && !filters.emailAssign && !filters.emailAssigned}
-              variant='outline'
-              onClick={resetFilters}
-              className='gap-2'
-            >
-              <FilterX className='h-4 w-4' />
-              Clear filters
-            </Button>
+            <AssetNameFilter
+              filters={filters}
+              setFilters={setFilters}
+            />
+            <AssignedByFilter
+              filters={filters}
+              setFilters={setFilters}
+            />
+            <AssignedToFilter
+              filters={filters}
+              setFilters={setFilters}
+            />
+            <ButtonClearFilter
+              filters={filters}
+              setFilters={setFilters}
+            />
           </div>
 
           {isLoading ? (
